backend/requirements.txt
fitz~=0.0.1.dev2
jupyter~=1.1.1
pip~=24.3.1
typing_extensions~=4.13.1
attrs~=25.3.0
filelock~=3.18.0
setuptools~=78.1.0
packaging~=24.2
platformdirs~=4.3.7
Pygments~=2.19.1
fastapi~=0.116.1
h11~=0.14.0
defusedxml~=0.7.1
ipython~=9.1.0
numpy~=2.2.3
pillow~=11.1.0
requests~=2.32.3
ipywidgets~=8.1.5
notebook~=7.4.0
PyYAML~=6.0.2
sniffio~=1.3.1
charset-normalizer~=3.4.1
pytz~=2025.1
overrides~=7.7.0
cffi~=1.17.1
regex~=2024.11.6
scipy~=1.15.2
threadpoolctl~=3.6.0
matplotlib~=3.10.1
decorator~=5.2.1
safetensors~=0.5.3
transformers~=4.51.3
dill~=0.3.8
tqdm~=4.67.1
psutil~=7.0.0
urllib3~=2.3.0
fsspec~=2025.3.0
aiohttp~=3.11.18
yarl~=1.20.0
MarkupSafe~=3.0.2
joblib~=1.4.2
certifi~=2025.7.9
pydantic~=2.11.7
pydantic_core~=2.33.2
starlette~=0.47.1
email_validator~=2.2.0
anyio~=4.9.0
scikit-learn~=1.6.1
pandas~=2.2.3
pyarrow~=20.0.0
datasets~=3.6.0
click~=8.1.8
uvicorn~=0.35.0
annotated-types~=0.7.0
idna~=3.10
Jinja2~=3.1.6
httpx~=0.28.1
tokenizers~=0.21.1
huggingface-hub~=0.31.1
nltk~=3.9.1
beautifulsoup4~=4.13.3
jupyterlab_pygments~=0.3.0
pathlib~=1.0.1
configparser~=7.2.0
pyOpenSSL~=25.1.0
cryptography~=45.0.7
torch~=2.2.2
python-docx~=1.2.0
PyMuPDF==1.23.7